+++ dreeve_polar_connector/CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 0.1.2
+
+- fix: pin the base image to the tag the upstream connector actually publishes
+
 ## 0.1.1
 
 - fix: use correct docker image tag
 
 ## 0.1.0
 
-- feat: bump Polar connector to v1.0.0 [Changelog](https://github.com/cgtobi/dreeve-polar-connector/releases)
+- feat: bump Polar connector to 0.1.0 [Changelog](https://github.com/cgtobi/dreeve-polar-connector/releases)